+++ source/count.js
@@ -3,7 +3,7 @@ import { equals } from './equals'
 
 export function count(searchFor, list){
   if (arguments.length === 1){
-    return listHolder => count(searchFor, listHolder)
+    return _list => count(searchFor, _list)
   }
   if (!_isArray(list)) return 0