+++ _ts-toolbelt/src/Boolean/_api.ts
@@ -1,9 +1,9 @@
 /** @ignore *//** */
 
-export {And} from './And'
-export {True, False, Boolean} from './Boolean'
-export {BooleanOf} from './BooleanOf'
-export {Format} from './Format'
-export {Not} from './Not'
-export {Or} from './Or'
-export {Xor} from './Xor'
+export type {And} from './And'
+export type {True, False, Boolean} from './Boolean'
+export type {BooleanOf} from './BooleanOf'
+export type {Format} from './Format'
+export type {Not} from './Not'
+export type {Or} from './Or'
+export type {Xor} from './Xor'