+++ _ts-toolbelt/src/Iteration/_api.ts
@@ -1,10 +1,10 @@
 /** @ignore *//** */
 
-export * as Maps from './Maps/_api'
-export {Format} from './Format'
-export {Iteration} from './Iteration'
-export {IterationOf} from './IterationOf'
-export {Key} from './Key'
-export {Next} from './Next'
-export {Pos} from './Pos'
-export {Prev} from './Prev'
+export type * as Maps from './Maps/_api'
+export type {Format} from './Format'
+export type {Iteration} from './Iteration'
+export type {IterationOf} from './IterationOf'
+export type {Key} from './Key'
+export type {Next} from './Next'
+export type {Pos} from './Pos'
+export type {Prev} from './Prev'