+++ _ts-toolbelt/src/Object/P/_api.ts
@@ -1,8 +1,8 @@
 /** @ignore *//** */
 
-export {Merge} from './Merge'
-export {Omit} from './Omit'
-export {Pick} from './Pick'
-export {Readonly} from './Readonly'
-export {Update} from './Update'
-export {Record} from './Record'
+export type {Merge} from './Merge'
+export type {Omit} from './Omit'
+export type {Pick} from './Pick'
+export type {Readonly} from './Readonly'
+export type {Update} from './Update'
+export type {Record} from './Record'