+++ _ts-toolbelt/src/String/_api.ts
@@ -1,4 +1,4 @@
 /** @ignore *//** */
 
-export {Format} from './Format'
-export {String} from './String'
+export type {Format} from './Format'
+export type {String} from './String'