+++ scripts/populate-readme-data/assets/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Add `R.union` method
 
-- `R.defaultTo`, `R.over` and `R.set` use `R.curry` instead of manual currying
+- `R.over` and `R.set` use `R.curry` instead of manual currying
 
 - `R.lensPath` typings support string as path, i.e. `'a.b'` instead of `['a', 'b']`